+++ IBll/User/Ipromotion_events_Bll.cs
@@ -31,7 +31,7 @@ namespace IBll
         #endregion
 
         #region 查询
-        List<promotion_coupons> GetPromotionCouponsList(string name, decimal value, int issue_status, DateTime? startdate1, DateTime? startdate2, DateTime? enddate1, DateTime? enddate2,
+        List<promotion_coupons> GetPromotionCouponsList(string name, int value, int issue_status, DateTime? startdate1, DateTime? startdate2, DateTime? enddate1, DateTime? enddate2,
            orderbyEnum? orderby, string orderbyCol, int pageIndex, int pageSize, out int record, out string error);
         #endregion
         #endregion
